Add smart-dupe script to tools
Finally figured out which version of 'smart-dupe' perl script I had
laying around was the most efficient, when I recently tested on
5000+ files.

I really wish I'd been tracking development of this script in git
previously, because it was impressive to see the changes I made over
time to go from a very long-running script to one that is awesomely
fast.

I am rather proud of this little guy.